+++ astropy/time/tests/test_basic.py
@@ -556,3 +556,7 @@ def test_TimeFormat_scale():
     t.delta_ut1_utc = 0.0
     t.unix
     assert t.unix == t.utc.unix
+
+def test_scale_conversion():
+    with pytest.raises(ScaleValueError):
+        t = Time(Time.now().cxcsec, format='cxcsec', scale='ut1')