+++ CHANGES.rst
@@ -311,6 +311,9 @@ Bug Fixes
 
 - ``astropy.time``
 
+  - Fix a problem where scale conversion problem in TimeFromEpoch
+    was not showing a useful error [#2046]
+
   - Fix a problem when converting to one of the formats `unix`,
     `cxcsec`, `gps` or `plot_date` when the time scale is `UT1`, `TDB`
     or `TCB` [#1732]